+++ gwcelery/tasks/detchar.py
@@ -152,6 +152,12 @@ def check_vector(channel, start, end, bitmask, logic_type):
 @app.task(shared=False)
 def check_vectors(event, superevent_id, start, end):
     """Perform data quality checks for an event."""
+    # Skip MDC events.
+    if event['search'] == 'MDC':
+        log.info('Skipping state vector checks because %s is an MDC',
+                 event['graceid'])
+        return
+
     instruments = event['instruments'].split(',')
 
     states = {key: check_vector(key, start, end, *value)